Prepare entities for database

+++ packages/server/src/entity/SectorWeighting.ts
@@ -23,6 +23,6 @@ export class SectorWeighting {
   @Column()
   sector: string
 
-  @Column({ type: 'decimal' })
+  @Column({ type: 'double' })
   percentage: number
 }
